+++ include/Cpp_Utils/Map.ipp
@@ -80,7 +80,7 @@ const Key & at_value(const std::map<Key, const Value> & map, const Value & value
 
 
 template<typename Key, typename Value>
-std::pair<Key, const Value> at_index(const std::map<Key, const Value> & map, int index)
+std::pair<Key, const Value> at_index(const std::map<Key, const Value> & map, size_t index)
 {
     if (index >= map.size())
     {
@@ -95,7 +95,7 @@ std::pair<Key, const Value> at_index(const std::map<Key, const Value> & map, int
 
 
 template<typename Key, typename Value>
-std::pair<Key, Value> at_index(const std::map<Key, Value> & map, int index)
+std::pair<Key, Value> at_index(const std::map<Key, Value> & map, size_t index)
 {
     if (index >= map.size())
     {
